1. 圖片連結加入 timestamp 看能否解決電腦版快取問題 2. 加入是否有迷宮的判斷

+++ routes/api.php
@@ -45,27 +45,27 @@ Route::post('lineBotWebhook', function (Request $request) {
     $bot = new LINEBot($httpClient, ['channelSecret' => env('LINE_BOT_CHANNEL_SECRET')]);
 
     // 取得 user
-    $response = $bot->getProfile($userId);
-    if ($response->isSucceeded()) {
-        $profile = $response->getJSONDecodedBody();
+    // $response = $bot->getProfile($userId);
+    // if ($response->isSucceeded()) {
+    //     $profile = $response->getJSONDecodedBody();
 
-        $lineUser = LineUser::find($userId);
-        if ($lineUser) {
-            $lineUser->update([
-                'userId'        => $profile['userId'],
-                'displayName'   => isset($profile['displayName']) ? $profile['displayName'] : '',
-                'pictureUrl'    => isset($profile['pictureUrl']) ? $profile['pictureUrl'] : '',
-                'statusMessage' => isset($profile['statusMessage']) ? $profile['statusMessage'] : '',
-            ]);
-        } else {
-            LineUser::create([
-                'userId'        => $profile['userId'],
-                'displayName'   => isset($profile['displayName']) ? $profile['displayName'] : '',
-                'pictureUrl'    => isset($profile['pictureUrl']) ? $profile['pictureUrl'] : '',
-                'statusMessage' => isset($profile['statusMessage']) ? $profile['statusMessage'] : '',
-            ]);
-        }
-    }
+    //     $lineUser = LineUser::find($userId);
+    //     if ($lineUser) {
+    //         $lineUser->update([
+    //             'userId'        => $profile['userId'],
+    //             'displayName'   => isset($profile['displayName']) ? $profile['displayName'] : '',
+    //             'pictureUrl'    => isset($profile['pictureUrl']) ? $profile['pictureUrl'] : '',
+    //             'statusMessage' => isset($profile['statusMessage']) ? $profile['statusMessage'] : '',
+    //         ]);
+    //     } else {
+    //         LineUser::create([
+    //             'userId'        => $profile['userId'],
+    //             'displayName'   => isset($profile['displayName']) ? $profile['displayName'] : '',
+    //             'pictureUrl'    => isset($profile['pictureUrl']) ? $profile['pictureUrl'] : '',
+    //             'statusMessage' => isset($profile['statusMessage']) ? $profile['statusMessage'] : '',
+    //         ]);
+    //     }
+    // }
 
     if ($message['type'] == 'text') {
         $text = $message['text'];
@@ -103,11 +103,19 @@ Route::post('lineBotWebhook', function (Request $request) {
             $textMessageBuilder = new TextMessageBuilder($replyMessageText);
             $response = $bot->replyMessage($replyToken, $textMessageBuilder);
         } elseif ($text == 'maze:show' || $text == 'maze:showAnswer') {
-            $imageMessageBuilder = new ImageMessageBuilder(
-                route('maze.show', [$targetType, $targetId, 0, ($text == 'maze:showAnswer') ? 1 : 0]),
-                route('maze.show', [$targetType, $targetId, 1, ($text == 'maze:showAnswer') ? 1 : 0])
-            );
-            $response = $bot->replyMessage($replyToken, $imageMessageBuilder);
+            $maze = Maze::where('targetType', $targetType)
+                    ->where('targetId', $targetId)
+                    ->first();
+            if ($maze) {
+                $imageMessageBuilder = new ImageMessageBuilder(
+                    route('maze.show', [$targetType, $targetId, 0, ($text == 'maze:showAnswer') ? 1 : 0, time()]),
+                    route('maze.show', [$targetType, $targetId, 1, ($text == 'maze:showAnswer') ? 1 : 0, time()])
+                );
+                $response = $bot->replyMessage($replyToken, $imageMessageBuilder);
+            } else {
+                $textMessageBuilder = new TextMessageBuilder("目前沒有迷宮喔！");
+                $response = $bot->replyMessage($replyToken, $textMessageBuilder);
+            }
         } elseif ($text == 'maze:close') {
             $maze = Maze::where('targetType', $targetType)
                     ->where('targetId', $targetId)
@@ -119,37 +127,40 @@ Route::post('lineBotWebhook', function (Request $request) {
             $maze = Maze::where('targetType', $targetType)
                     ->where('targetId', $targetId)
                     ->first();
-            $mazeData = unserialize($maze->mazeData);
-            $newLocation = $maze->location;
 
-            if (strtolower($text) == 'w' && in_array($mazeData[$maze->location], [1,3,5,7,9,11,13,15,16])) {
-                $direction = '上';
-                $newLocation = $maze->location - $maze->width;
-            } elseif (strtolower($text) == 'a' && in_array($mazeData[$maze->location], [8,9,10,11,12,13,14,15,16])) {
-                $direction = '左';
-                $newLocation = $maze->location - 1;
-            } elseif (strtolower($text) == 's' && in_array($mazeData[$maze->location], [4,5,6,7,12,13,14,15,16])) {
-                $direction = '下';
-                $newLocation = $maze->location + $maze->width;
-            } elseif (strtolower($text) == 'd' && in_array($mazeData[$maze->location], [2,3,6,7,10,11,14,15,16])) {
-                $direction = '右';
-                $newLocation = $maze->location + 1;
-            } else {
-                $textMessageBuilder = new TextMessageBuilder('此路不通');
+            if ($maze) {
+                $mazeData = unserialize($maze->mazeData);
+                $newLocation = $maze->location;
+
+                if (strtolower($text) == 'w' && in_array($mazeData[$maze->location], [1,3,5,7,9,11,13,15,16])) {
+                    $direction = '上';
+                    $newLocation = $maze->location - $maze->width;
+                } elseif (strtolower($text) == 'a' && in_array($mazeData[$maze->location], [8,9,10,11,12,13,14,15,16])) {
+                    $direction = '左';
+                    $newLocation = $maze->location - 1;
+                } elseif (strtolower($text) == 's' && in_array($mazeData[$maze->location], [4,5,6,7,12,13,14,15,16])) {
+                    $direction = '下';
+                    $newLocation = $maze->location + $maze->width;
+                } elseif (strtolower($text) == 'd' && in_array($mazeData[$maze->location], [2,3,6,7,10,11,14,15,16])) {
+                    $direction = '右';
+                    $newLocation = $maze->location + 1;
+                } else {
+                    $textMessageBuilder = new TextMessageBuilder('此路不通');
+                }
+
+                if ($newLocation == 24) {
+                    $textMessageBuilder = new TextMessageBuilder('恭喜你到達終點啦！');
+                    $maze->delete();
+                } elseif ($newLocation != $maze->location) {
+                    $maze->update([
+                        'location' => $newLocation,
+                    ]);
+
+                    $textMessageBuilder = new TextMessageBuilder("你往{$direction}走了一步，現在位置：$newLocation");
+                }
+
+                $response = $bot->replyMessage($replyToken, $textMessageBuilder);
             }
-
-            if ($newLocation == 24) {
-                $textMessageBuilder = new TextMessageBuilder('恭喜你到達終點啦！');
-                $maze->delete();
-            } elseif ($newLocation != $maze->location) {
-                $maze->update([
-                    'location' => $newLocation,
-                ]);
-
-                $textMessageBuilder = new TextMessageBuilder("你往{$direction}走了一步，現在位置：$newLocation");
-            }
-
-            $response = $bot->replyMessage($replyToken, $textMessageBuilder);
         }
     }
